moving interrupt test to a separate file
