+++ week07/manager_test.go
@@ -2,10 +2,7 @@ package week07
 
 import (
 	"context"
-	"os/signal"
-	"runtime"
 	"sync"
-	"syscall"
 	"testing"
 	"time"
 )
@@ -23,40 +20,6 @@ func Test_Run(t *testing.T) {
 		products := []*Product{
 			// 6_000 milliseconds = 6 seconds
 			&Product{Quantity: 6_000},
-		}
-
-		cprod, err := Run(ctx, 1, products...)
-		if err != nil {
-			t.Errorf("got error %v, want nil", err)
-		}
-
-		if len(cprod) != 0 {
-			t.Errorf("got %v, want 0", len(cprod))
-		}
-	})
-
-	t.Run("interruption by a signal", func(t *testing.T) {
-		if runtime.GOOS == "windows" {
-			t.Skip("this test can only run on unix systems due to the signal")
-		}
-
-		t.Parallel()
-
-		ctx, cnFn := signal.NotifyContext(context.Background(), syscall.SIGUSR2)
-		defer cnFn()
-
-		go func() {
-			time.Sleep(time.Second)
-
-			t.Log("sending test signal")
-
-			// send the TEST_SIGNAL to the system
-			syscall.Kill(syscall.Getpid(), syscall.SIGUSR2)
-		}()
-
-		products := []*Product{
-			// 1_500 milliseconds = 1.5 seconds
-			&Product{Quantity: 1_500},
 		}
 
 		cprod, err := Run(ctx, 1, products...)
